requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
sqlparse==0.5.0
psycopg2==2.9.9
drf-spectacular==0.27.2
pillow==10.4.0